Complete refactoring, seperating out OCR methods from collection-facing methods
Built up a heirachy of Notes Query which contains OCRNote(s), which contain OCRField(s), which contain OCRImage(s), allowing the passing around of references to these objects. Enables easy multithreading of OCRImages with the referneces objects then updated and written to the Collection database.

Implments ThreadPoolExecutor as suggested by @phu54321 in #7